+++ src/StackTrace.php
@@ -5,6 +5,7 @@ namespace Awesomite\StackTrace;
 use Awesomite\Iterators\CallbackIterator;
 use Awesomite\StackTrace\Arguments\Values\DeserializedValue;
 use Awesomite\StackTrace\Arguments\Values\Value;
+use Awesomite\StackTrace\Exceptions\LogicException;
 use Awesomite\StackTrace\SourceCode\File;
 use Awesomite\StackTrace\Steps\Step;
 use Awesomite\StackTrace\Steps\StepInterface;
@@ -17,8 +18,8 @@ use Composer\Semver\Semver;
  */
 class StackTrace implements StackTraceInterface
 {
-    const VERSION = '0.9.2';
-    const CONSTRAINTS_VERSION = '>=0.1.0 <0.10.0';
+    const VERSION = '0.10.0';
+    const CONSTRAINTS_VERSION = '>=0.1.0 <0.11.0';
 
     private $arrayStackTrace;
 
@@ -92,7 +93,7 @@ class StackTrace implements StackTraceInterface
         $data = unserialize($serialized);
         if (!Semver::satisfies($data['__version'], static::CONSTRAINTS_VERSION)) {
             $message = 'Cannot use incompatible version to unserialize stack trace (serialized by: %s, current: %s).';
-            throw new \LogicException(sprintf($message, $data['__version'], static::VERSION));
+            throw new LogicException(sprintf($message, $data['__version'], static::VERSION));
         }
         $this->arrayStackTrace = $data['steps'];
         $this->unserialized = true;
@@ -160,32 +161,38 @@ class StackTrace implements StackTraceInterface
      */
     public function convertStep(array $step, $toSerialize = false)
     {
-        $result = array();
-        foreach ($step as $key => $value) {
-            $result[$key] = $value;
-        }
-
         if ($this->withoutArgs) {
-            $result['args'] = array();
-        } else if (empty($result[Constants::KEY_ARGS_CONVERTED]) && isset($result['args'])) {
-            $result['args'] = $this->convertArgs($result['args']);
-            $result[Constants::KEY_ARGS_CONVERTED] = true;
+            $step['args'] = array();
+        } else if (empty($step[Constants::KEY_ARGS_CONVERTED]) && isset($step['args'])) {
+            $maxArgs = null;
+            if (version_compare(PHP_VERSION, '5.6') >= 0) {
+                $fakeStep = new Step($step);
+                $reflectionFn = $fakeStep->hasCalledFunction() && $fakeStep->getCalledFunction()->hasReflection()
+                    ? $fakeStep->getCalledFunction()->getReflection()
+                    : null;
+                if (!is_null($reflectionFn) && $reflectionFn->isVariadic()) {
+                    $maxArgs = count($reflectionFn->getParameters());
+                }
+            }
+
+            $step['args'] = $this->convertArgs($step['args'], $maxArgs);
+            $step[Constants::KEY_ARGS_CONVERTED] = true;
         }
 
         if (
             !$toSerialize
-            && !isset($result[Constants::KEY_FILE_OBJECT])
-            && isset($result['file'])
-            && isset($this->files[$result['file']])
+            && !isset($step[Constants::KEY_FILE_OBJECT])
+            && isset($step['file'])
+            && isset($this->files[$step['file']])
         ) {
-            $result[Constants::KEY_FILE_OBJECT] = $this->files[$result['file']];
+            $step[Constants::KEY_FILE_OBJECT] = $this->files[$step['file']];
         }
 
-        if (isset($result['object'])) {
-            unset($result['object']);
+        if (isset($step['object'])) {
+            unset($step['object']);
         }
 
-        return $result;
+        return $step;
     }
 
     private function getVarDumper()
@@ -197,19 +204,32 @@ class StackTrace implements StackTraceInterface
         return $this->varDumper;
     }
 
-    private function convertArgs(array $args)
+    /**
+     * @param array $inputArgs
+     * @param int|null $maxArgs
+     * @return array
+     */
+    private function convertArgs(array $inputArgs, $maxArgs)
     {
         /**
-         * input has to be copied to different array,
-         * because array $args returned by debug_backtrace function contains references from PHP 7.0
+         * debug_backtrace()[$x]['args'] can contain references
          */
-        $result = array();
-
-        foreach ($args as $key => $value) {
-            $result[$key] = $this->convertArg($value);
+        $args = array();
+        foreach ($inputArgs as $key => $value) {
+            $args[$key] = $value;
         }
 
-        return $result;
+        if (!is_null($maxArgs) && $maxArgs <= count($args)) {
+            $preparedCopy = $args;
+            $args = array_slice($preparedCopy, 0, $maxArgs - 1);
+            $args[] = array_slice($preparedCopy, $maxArgs - 1);
+        }
+
+        foreach ($args as $key => $value) {
+            $args[$key] = $this->convertArg($value);
+        }
+
+        return $args;
     }
 
     private function convertArg($value)